Makefile: add libraries to deb
This is probably a bad way to do this. But I can't use the Debian
packages shipped with the operating system because:

1. Scalpl is not even in the repositories (I don't blame Debian, it's
a new library)
2. Construct is very old and hasn't been updated because the update
breaks existing applications depending on it.

I therefore had to ship the versions I use in the deb. This is
probably kind of ugly, but I don't know of a better way yet.