Use Int24ul as the Luc data type
Per inspection of the save ram and the FAQ mentioned in the credits,
Luc is stored as a 24bit integer at offsets 491-493. Using an Int16ul
limited Luc in the app to ~65k.

Section: 3.o
https://gamefaqs.gamespot.com/snes/588648-seiken-densetsu-3/faqs/9788

Python core type def comment, heh
"""A 3-byte little-endian unsigned integer, as used in ancient file formats."""