Read location values as 12 bit integers
Based on research from issue #2, the game appears to only use 12 bits
of the 16 from offsets 726-727. The data in the leading four bits of
the most significant byte does not appear to affect loading location
at all.

The following two values loaded in the same location in game:

  dd 42 = 1101 1101 >0100< 0010
  dd 02 = 1101 1101 >0000< 0010

On load, the app now ignores those four bits as the game apparently
does, and on write they are zeroed out.